fastapi-backend/requirements.txt
# requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
pymysql
cryptography
starlette
PyJWT
bcrypt
python-dotenv
httpx
resend